Moved UI related stuff out of `ArchiveManager`
