+++ WolvenKit.Modkit/RED4/Tasks/UnbundleTask.cs
@@ -34,7 +34,7 @@ public partial class ConsoleFunctions
         if (!string.IsNullOrEmpty(options.gamepath) && Directory.Exists(options.gamepath))
         {
             var exePath = new FileInfo(Path.Combine(options.gamepath, "bin", "x64", "Cyberpunk2077.exe"));
-            _archiveManager.LoadGameArchives(exePath, false);
+            _archiveManager.LoadGameArchives(exePath);
         }
 
         var result = 0;
